requirements.txt
snowflake-snowpark-python==1.20.0
streamlit==1.37.0
plotly==5.23.0